Extracting collaborators to allow easier integration in legacy applications
